requirements.txt
streamlit
python-dotenv
pydantic
numpy
pandas
tqdm

pypdf
pdfplumber

rank-bm25
faiss-cpu
sentence-transformers

openai

pytesseract
pillow

pytest
ruff
